requirements.txt
aiohttp==3.9.1
deepgram-sdk==3.1.0
elevenlabs==0.2.27
fastapi==0.108.0
itsdangerous==2.1.2
langchain==0.0.353
langchain-community==0.0.7
litellm==1.16.8
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
starlette==0.32.0.post1
setuptools
twilio==8.11.0
uvicorn==0.25.0
websockets==12.0
boto3>=1.34.72
langchain-aws==0.1.0
streamlit==1.28.0
pandas==2.1.3
